ffi: cleanup pystate
