+++ src/ffi/cpython/mod.rs
@@ -2,6 +2,7 @@ pub(crate) mod abstract_;
 // skipped bytearrayobject.h
 #[cfg(not(PyPy))]
 pub(crate) mod bytesobject;
+#[cfg(not(PyPy))]
 pub(crate) mod ceval;
 pub(crate) mod code;
 pub(crate) mod compile;
@@ -18,13 +19,12 @@ pub(crate) mod object;
 pub(crate) mod pydebug;
 #[cfg(all(Py_3_8, not(PyPy)))]
 pub(crate) mod pylifecycle;
-
-#[cfg(all(Py_3_8, not(PyPy)))]
 pub(crate) mod pystate;
 
 pub use self::abstract_::*;
 #[cfg(not(PyPy))]
 pub use self::bytesobject::*;
+#[cfg(not(PyPy))]
 pub use self::ceval::*;
 pub use self::code::*;
 pub use self::compile::*;
@@ -39,6 +39,4 @@ pub use self::object::*;
 pub use self::pydebug::*;
 #[cfg(all(Py_3_8, not(PyPy)))]
 pub use self::pylifecycle::*;
-
-#[cfg(all(Py_3_8, not(PyPy)))]
 pub use self::pystate::*;
