+++ src/ffi/pystate.rs
@@ -1,26 +1,49 @@
+#[cfg(not(PyPy))]
 use crate::ffi::moduleobject::PyModuleDef;
 use crate::ffi::object::PyObject;
-use crate::ffi::PyFrameObject;
-use std::os::raw::{c_int, c_long};
+use std::os::raw::c_int;
+#[cfg(not(PyPy))]
+use std::os::raw::c_long;
 
 pub const MAX_CO_EXTRA_USERS: c_int = 255;
 
+opaque_struct!(PyThreadState);
 opaque_struct!(PyInterpreterState);
 
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub struct PyThreadState {
-    pub ob_base: PyObject,
-    pub interp: *mut PyInterpreterState,
-}
-
 extern "C" {
+    #[cfg(not(PyPy))]
+    #[cfg_attr(docsrs, doc(cfg(not(PyPy))))]
     pub fn PyInterpreterState_New() -> *mut PyInterpreterState;
+    #[cfg(not(PyPy))]
+    #[cfg_attr(docsrs, doc(cfg(not(PyPy))))]
     pub fn PyInterpreterState_Clear(arg1: *mut PyInterpreterState);
+    #[cfg(not(PyPy))]
+    #[cfg_attr(docsrs, doc(cfg(not(PyPy))))]
     pub fn PyInterpreterState_Delete(arg1: *mut PyInterpreterState);
-    //fn _PyState_AddModule(arg1: *mut PyObject,
-    //                      arg2: *mut PyModuleDef) -> c_int;
+
+    #[cfg(all(Py_3_9, not(PyPy)))]
+    #[cfg_attr(docsrs, doc(all(Py_3_9, not(PyPy))))]
+    pub fn PyInterpreterState_Get() -> *mut PyInterpreterState;
+
+    #[cfg(all(Py_3_8, not(PyPy)))]
+    #[cfg_attr(docsrs, doc(all(Py_3_8, not(PyPy))))]
+    pub fn PyInterpreterState_GetDict() -> *mut PyObject;
+
+    #[cfg(all(Py_3_7, not(PyPy)))]
+    #[cfg_attr(docsrs, doc(all(Py_3_7, not(PyPy))))]
+    pub fn PyInterpreterState_GetID() -> i64;
+
+    #[cfg(not(PyPy))]
+    #[cfg_attr(docsrs, doc(cfg(not(PyPy))))]
+    pub fn PyState_AddModule(arg1: *mut PyObject, arg2: *mut PyModuleDef) -> c_int;
+    #[cfg(not(PyPy))]
+    #[cfg_attr(docsrs, doc(cfg(not(PyPy))))]
+    pub fn PyState_RemoveModule(arg1: *mut PyModuleDef) -> c_int;
+
+    #[cfg(not(PyPy))]
+    #[cfg_attr(docsrs, doc(cfg(not(PyPy))))]
     pub fn PyState_FindModule(arg1: *mut PyModuleDef) -> *mut PyObject;
+
     #[cfg_attr(PyPy, link_name = "PyPyThreadState_New")]
     pub fn PyThreadState_New(arg1: *mut PyInterpreterState) -> *mut PyThreadState;
     //fn _PyThreadState_Prealloc(arg1: *mut PyInterpreterState)
@@ -39,6 +62,8 @@ extern "C" {
     pub fn PyThreadState_Swap(arg1: *mut PyThreadState) -> *mut PyThreadState;
     #[cfg_attr(PyPy, link_name = "PyPyThreadState_GetDict")]
     pub fn PyThreadState_GetDict() -> *mut PyObject;
+    #[cfg(not(PyPy))]
+    #[cfg_attr(docsrs, doc(cfg(not(PyPy))))]
     pub fn PyThreadState_SetAsyncExc(arg1: c_long, arg2: *mut PyObject) -> c_int;
 }
 
@@ -54,18 +79,12 @@ extern "C" {
     pub fn PyGILState_Ensure() -> PyGILState_STATE;
     #[cfg_attr(PyPy, link_name = "PyPyGILState_Release")]
     pub fn PyGILState_Release(arg1: PyGILState_STATE);
+    #[cfg(not(PyPy))]
+    #[cfg_attr(docsrs, doc(cfg(not(PyPy))))]
     pub fn PyGILState_GetThisThreadState() -> *mut PyThreadState;
-    pub fn PyGILState_Check() -> c_int;
 }
 
 #[inline]
 pub unsafe fn PyThreadState_GET() -> *mut PyThreadState {
     PyThreadState_Get()
 }
-
-pub type Py_tracefunc = extern "C" fn(
-    obj: *mut PyObject,
-    frame: *mut PyFrameObject,
-    what: c_int,
-    arg: *mut PyObject,
-) -> c_int;
